Move two play sound methods from PositionUtils to WorldUtils and delete PositionUtils
